inline stringable check & cast
